Split out parser tests into a separate file
